+++ frontend/wailsjs/go/app/App.d.ts
@@ -213,6 +213,8 @@ export function GetAllAccountIdentities():Promise<Array<app.AccountIdentityGroup
 
 export function GetAlwaysLoadImages():Promise<boolean>;
 
+export function GetAlwaysShowMessageCheckbox():Promise<boolean>;
+
 export function GetAppInfo():Promise<app.AppInfo>;
 
 export function GetAttachment(arg1:string):Promise<message.Attachment>;
@@ -548,6 +550,8 @@ export function SetAccountEnabled(arg1:string,arg2:boolean):Promise<void>;
 export function SetAddressbookEnabled(arg1:string,arg2:boolean):Promise<void>;
 
 export function SetAlwaysLoadImages(arg1:boolean):Promise<void>;
+
+export function SetAlwaysShowMessageCheckbox(arg1:boolean):Promise<void>;
 
 export function SetAutostart(arg1:boolean):Promise<void>;